test(store): cover every released build as a migration source
The SQLite layout has not shipped, so every install in the field is on Hive
and 1466, 1480 and 1491 are all upgrade sources. The suite ran against one.

`hive_adapters.g.dart` is byte-identical across the three, so they share one
set of assertions — but that is a fact worth checking rather than assuming,
so 1480 gets a fixture generated from its own tag even though the generator
ran against it unchanged. 1491 shipped an `agent_conversation` box the
others do not have, which is the case that needed its own data.

The test is now parameterised over the three. `Paths.doc` is `late final`
and cannot be set per group, so one temp directory is refilled from the
fixture under test in `setUp`.

Building the 1491 conversations by hand first is what caught the reason
these fixtures exist: written as JSON with camelCase keys and a `type`
discriminator they were silently dropped on import, because the release
writes snake_case and `kind`. They are built through 1491's own model and
serialiser now, and the README says why.
