+++ test/fixtures/hive_v1480/conn_stats_index.hive
@@ -1,0 +1,2 @@
+srv-pwd
+srv-key
